refactor: products.ts
