--- product-service/lambda-functions/product.interface.ts
@@ -1,9 +1,0 @@
-export interface IProduct {
-  id: string;
-  title: string;
-  description: string;
-  price: number;
-  count: number;
-}
-
-export type ProductInfo = Omit<IProduct, "id">;
